+++ changelog.md
@@ -12,7 +12,8 @@
 + The container that holds the navigation buttons now has the class 'stepy-buttons';
 + The error container class was renamed to plural 'stepy-errors';
 + Forms without ID will receives a genarated hash to be used as hook;
-+ The 'stepy-titles' class was renamed to 'stepy-header'.
++ The 'stepy-titles' class was renamed to 'stepy-header';
++ The 'stepy-{id}-titles' ID was renamed to 'stepy-{id || hash}-header'.
 
 ## 1.1.0